requirements.txt
# Core dependencies
streamlit>=1.29.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
PyYAML>=6.0

# Document processing
PyMuPDF>=1.23.8
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.1.0

# Arabic NLP
# CAMeL-Tools==1.5.4
pyarabic>=0.6.15
farasapy>=0.0.14

# Vector database and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2
langchain>=0.0.340
langchain-community>=0.0.1
faiss-cpu>=1.7.4

# LLM integration
ollama>=0.1.7
transformers>=4.36.0
openai>=1.12.0
# torch==2.1.0

# Utilities
pandas>=2.1.3
numpy>=1.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1
